Data.Set.Symbol

Description

Type-level Sets of Symbols.

While it's possible to use this module without TypeApplications, it is recommended.

Data types

newtype Set (a :: [Symbol])

Type-level sets:

 λ> set :: Set '["x", "y", "z"]
 Set ["x","y","z"]

 λ> set :: Set '["x", "z", "y"]
 interactive:142:1: error:
     • Couldn't match type ‘'GT’ with ‘'LT’ arising from a use of ‘set’
     • In the expression: set :: Set '["x", "z", "y"]
       In an equation for ‘it’: it = set :: Set '["x", "z", "y"]

Or with TypeApplications (much more concise):

 λ> :t set @["x", "y", "z"]
 set @["x", "y", "z"] :: Set '["x", "y", "z"]

Constructors

Set

Fields

setDict :: Dict (IsSet a)

Core classes

type family IsSet (a :: [Symbol]) = (result :: Constraint) | result -> a where ...

This is an injective type-family that asserts that a list of Symbols represent a valid Set.

Equations

IsSet '[] = ()

IsSet (x ': xs) = (KnownSymbol x, IsSet1 x xs, IsSet xs)

type family IsSet1 (a :: Symbol) (as :: [Symbol]) = (result :: Constraint) | result -> a as where ...

Helper for IsSet, assuming a non-empty set by requiring an additional parameter.

Equations

IsSet1 x '[] = KnownSymbol x

IsSet1 x (y ': ys) = (KnownSymbol x, CmpSymbol x y ~ LT, IsSet (y ': ys))

class ReifySet (a :: [Symbol]) where

The class of lists of symbols that can be reflected from a Set

Minimal complete definition

reflectSet

Methods

reflectSet :: Set a -> [String]

emptyIsSet :: Set '[]

The empty Set is a Set

insertMinIsSet :: (KnownSymbol x, IsSet1 x xs) => proxy x -> Set xs -> Set (x ': xs)

Insert a new KnownSymbol, that's known to be the minimum of the Set

insertMinIsSet1 :: (KnownSymbol x, CmpSymbol x y ~ LT) => proxy x -> Set1 y ys -> Set1 x (y ': ys)

We have a bijection between Min1 and Set1 when we have AllKnownSym1 fromMin1 :: (AllKnownSym1 x xs, Min1 x xs, Set xs) :- Set1 x xs fromMin1 =

Insert a new minimum KnownSymbol into a Set1

tailSetIsSet :: Set (x ': xs) -> Set xs

The tail of a Set is a Set

tailSet1IsSet :: Set1 x xs -> Set xs

The tail of a Set1 is a Set

newtype Set1 (a :: Symbol) (as :: [Symbol])

Non-empty Set

Constructors

Set1

Fields

set1Dict :: Dict (IsSet (a ': as))

set1 :: IsSet (a ': as) => Set1 a as

If we know a ': as represents a Set, we can instantiate a Set1 a as

emptyKnownIsSet1 :: KnownSymbol x => Set1 x '[]

If x is a KnownSymbol, we can form a singleton Set1

newtype AllKnownSym1 (a :: Symbol) (as :: [Symbol])

A proof that all the Symbols in the type-level list, as well as the first provided Symbol, are all KnownSymbols

Constructors

AllKnownSym1

Fields

allKnownSym1Dict :: Dict (IsAllKnownSym1 a as)

newtype Min1 (a :: Symbol) (as :: [Symbol])

Min1 encodes a non-empty AllKnownSym1 Set1

Constructors

Min1

Fields

min1Dict :: Dict (IsMin a as)

Construction and first operatora

set :: IsSet a => Set a

Construct a Set using its type, e.g.

 λ> set :: Set '["x", "y", "z"]
 Set ["x","y","z"]

Or with TypeApplications (much more concise):

 λ> set @["x", "y", "z"]
 Set ["x", "y", "z"]

tailIsSet :: IsSet (a ': as) :- IsSet as

A proof that the tail of a Set is a Set

headIsKnown :: IsSet (x ': xs) :- KnownSymbol x

A proof that the head of a Set is a KnownSymbol

headProxy :: Set (a ': as) -> Proxy a

Type-level head

tail :: Set (a ': as) -> Set as

Extract the elements after the head of a Set, which is non-empty

unconsProxy :: Set (a ': as) -> (Proxy a, Set as)

Type-level uncons

head :: Set (a ': as) -> String

Extract the first (least) element of a list, which is non-empty.

uncons :: Set (a ': as) -> (String, Set as)

Decompose a Set into its head and tail.

append :: IsSet (a :<|> b) => Set a -> Set b -> Set (a :<|> b)

Append two sets

 λ> set  '["x", "y"] `append` set  '["z"]
 Set ["x","y","z"]

type family (a :: [Symbol]) :<|> (b :: [Symbol]) = (result :: [Symbol]) where ... infixl 3

Append two Symbol lists that represent Sets

Equations

'[] :<|> y = y

x :<|> '[] = x

(x ': xs) :<|> ys = x :<| (xs :<|> ys)

type family (a :: Symbol) :<| (b :: [Symbol]) :: [Symbol] where ... infixl 2

Insert a Symbol into a Symbol list of a Set

Equations

x :<| '[] = '[x]

x :<| (y ': ys) = ConsHelper (CmpSymbol x y) x y ys

type family ConsHelper (c :: Ordering) (a :: Symbol) (b :: Symbol) (bs :: [Symbol]) :: [Symbol] where ...

a b result -> c, a b result -> bs, a c result -> b

Equations

ConsHelper LT x y ys = x ': (y ': ys)

ConsHelper EQ x y ys = y ': ys

ConsHelper GT x y ys = y ': (x :<| ys)

elem :: Reifies (ElemList s a) Bool => proxy (s :: Symbol) -> Set a -> Bool

Is a Symbol an elem of the given Set?

 λ> elem (Proxy "y") (set  '["x","y","z"])
 True

 λ> elem (Proxy "y") (set  '["x","z"])
 False

type family ElemList (a :: Symbol) (as :: [Symbol]) :: Bool where ...

The type family behind elem

Equations

ElemList _ '[] = False

ElemList x (y ': ys) = ElemListHelper (CmpSymbol x y) x y ys

type family ElemListHelper (c :: Ordering) (a :: Symbol) (b :: Symbol) (as :: [Symbol]) :: Bool where ...

ElemList assuming non-empty and with known Ordering

Equations

ElemListHelper GT x _ ys = ElemList x ys

ElemListHelper EQ _ _ _ = True

ElemListHelper LT _ _ _ = False

Operators

(\\) :: IsSet (Difference a b) => Set a -> Set b -> Set (Difference a b)

Alias for difference

Query

null :: Reifies (a == '[]) Bool => Set a -> Bool

Is this the empty Set?

nullProxy :: Set a -> Proxy (a == '[])

Type-level implementation of null

size :: KnownNat (Size a) => Set a -> Natural

The number of elements in the Set.

sizeProxy :: Set a -> Proxy (Size a)

Type-level size

type family Size (as :: [Symbol]) :: Nat where ...

Type family encoding of size

Equations

Size '[] = 0

Size (_ ': as) = 1 + Size as

member :: Reifies (ElemList s a) Bool => proxy (s :: Symbol) -> Set a -> Bool

Is the element in the Set?

notMember :: Reifies (ElemList s a) Bool => proxy (s :: Symbol) -> Set a -> Bool

Is the element not in the Set?

lookupLT :: Reifies (LookupLT a as) (Maybe String) => proxy (a :: Symbol) -> Set as -> Maybe String

Find largest element smaller than the given one.

 lookupLT 3 (fromList [3, 5]) == Nothing
 lookupLT 5 (fromList [3, 5]) == Just 3

lookupLTProxy :: proxy (a :: Symbol) -> Set as -> Proxy (LookupLT a as :: Maybe Symbol)

Type-level lookupLT

type family LookupLT (a :: Symbol) (as :: [Symbol]) :: Maybe Symbol where ...

Type family encoding of lookupLT

Equations

LookupLT _ '[] = Nothing

LookupLT x '[y] = JustIf y (CmpSymbol x y == LT)

LookupLT x (y ': (z ': ws)) = (((CmpSymbol x y == LT) && (CmpSymbol x z /= LT)) :? LookupLT x (z ': ws)) (Just y)

type family ((a :: Bool) :? (b :: k)) (c :: k) :: k where ...

Type family encoding of the ternary operator (if False then first argument else second)

Equations

(False :? x) _ = x

(True :? _) y = y

type family JustIf (a :: Symbol) (b :: Bool) :: Maybe Symbol where ...

Type family encoding of:

 x b ->
   if b
     then Just x
     else Nothing

Equations

JustIf x b = (b :? Nothing) (Just x)

lookupGT :: Reifies (LookupGT a as) (Maybe String) => proxy (a :: Symbol) -> Set as -> Maybe String

Find smallest element greater than the given one.

 lookupGT 4 (fromList [3, 5]) == Just 5
 lookupGT 5 (fromList [3, 5]) == Nothing

lookupGTProxy :: proxy (a :: Symbol) -> Set as -> Proxy (LookupGT a as :: Maybe Symbol)

Type-level lookupGT

type family LookupGT (a :: Symbol) (as :: [Symbol]) :: Maybe Symbol where ...

Type family encoding of lookupGT

Equations

LookupGT _ '[] = Nothing

LookupGT x (y ': ys) = ((CmpSymbol x y == LT) :? LookupGT x ys) (Just y)

lookupLE :: Reifies (LookupLE a as) (Maybe String) => proxy (a :: Symbol) -> Set as -> Maybe String

Find largest element smaller or equal to the given one.

 lookupLE 2 (fromList [3, 5]) == Nothing
 lookupLE 4 (fromList [3, 5]) == Just 3
 lookupLE 5 (fromList [3, 5]) == Just 5

lookupLEProxy :: proxy (a :: Symbol) -> Set as -> Proxy (LookupLE a as :: Maybe Symbol)

Type-level lookupLE

type family LookupLE (a :: Symbol) (as :: [Symbol]) :: Maybe Symbol where ...

Type family encoding of lookupLE

Equations

LookupLE _ '[] = Nothing

LookupLE x '[y] = JustIf y (CmpSymbol y x /= GT)

LookupLE x (y ': (z ': ws)) = (((CmpSymbol y x /= GT) && (CmpSymbol x z == LT)) :? LookupLE x (z ': ws)) (Just y)

lookupGE :: Reifies (LookupGE a as) (Maybe String) => proxy (a :: Symbol) -> Set as -> Maybe String

Find smallest element greater or equal to the given one.

 lookupGE 3 (fromList [3, 5]) == Just 3
 lookupGE 4 (fromList [3, 5]) == Just 5
 lookupGE 6 (fromList [3, 5]) == Nothing

lookupGEProxy :: proxy (a :: Symbol) -> Set as -> Proxy (LookupGE a as :: Maybe Symbol)

Type-level lookupGE

type family LookupGE (a :: Symbol) (as :: [Symbol]) :: Maybe Symbol where ...

Type family encoding of LookupGE

Equations

LookupGE _ '[] = Nothing

LookupGE x (y ': ys) = ((CmpSymbol x y /= GT) :? LookupGE x ys) (Just y)

isSubsetOf :: Reifies (IsSubsetOf a b) Bool => Set a -> Set b -> Bool

Is this a subset? (s1 isSubsetOf s2) tells whether s1 is a subset of s2@.

isSubsetOfProxy :: Set a -> Set b -> Proxy (IsSubsetOf a b :: Bool)

Type-level isSubsetOf

type family IsSubsetOf (a :: [Symbol]) (b :: [Symbol]) :: Bool where ...

Type family encoding of isSubsetOf

Equations

IsSubsetOf '[] _ = True

IsSubsetOf (x ': xs) ys = ElemList x ys && IsSubsetOf xs ys

isProperSubsetOf :: Reifies (IsSubsetOf a b && Not (a == b)) Bool => Set a -> Set b -> Bool

Is this a proper subset? (ie. a subset but not equal).

isProperSubsetOfProxy :: Set a -> Set b -> Proxy (IsProperSubsetOf a b :: Bool)

Type-level isProperSubsetOf

type family IsProperSubsetOf (a :: [Symbol]) (b :: [Symbol]) :: Bool where ...

Type family encoding of isProperSubsetOf

Equations

IsProperSubsetOf xs ys = IsSubsetOf xs ys && Not (xs == ys)

disjoint :: Reifies (Intersection a b == '[]) Bool => Set a -> Set b -> Bool

Check whether two Sets are disjoint (i.e. their intersection is empty).

 disjoint (fromList [2,4,6])   (fromList [1,3])     == True
 disjoint (fromList [2,4,6,8]) (fromList [2,3,5,7]) == False
 disjoint (fromList [1,2])     (fromList [1,2,3,4]) == False
 disjoint (fromList [])        (fromList [])        == True

disjointProxy :: Set a -> Set b -> Proxy (Disjoint a b :: Bool)

Type-level disjoint

type family Disjoint (a :: [Symbol]) (b :: [Symbol]) :: Bool where ...

Type family encoding of disjoint

Equations

Disjoint xs ys = Intersection xs ys == '[]

Construction

empty :: Set '[]

The empty Set.

singleton :: KnownSymbol a => proxy (a :: Symbol) -> Set '[a]

Construct a Set containing a single element

insert :: IsSet (a ': as) => proxy (a :: Symbol) -> Set as -> Set (a ': as)

Insert an element in a set. If the set already contains an element equal to the given value, it is replaced with the new value.

delete :: IsSet (Delete a as) => proxy (a :: Symbol) -> Set as -> Set (Delete a as)

Delete an element from a set.

type family Delete (a :: Symbol) (as :: [Symbol]) :: [Symbol] where ...

Type family encoding of delete

Equations

Delete _ '[] = '[]

Delete x (y ': ys) = ((x == y) :? (y ': Delete x ys)) ys

powerSet :: IsSet (PowerSet a) => Set a -> Set (PowerSet a)

We do:

 powerset someSet = Set [concat subset | subset <- toList (powerSet someSet)]

type family PowerSet (a :: [Symbol]) = (result :: [Symbol]) where ...

Type family encoding of powerSet

Equations

PowerSet '[] = '[]

PowerSet (x ': xs) = PowerSet xs :<|> PowerSet (Prepend x xs)

type family Prepend (a :: Symbol) (as :: [Symbol]) = (result :: [Symbol]) where ...

Prepend the given Symbol to all elements of the list (preserves order)

Equations

Prepend _ '[] = '[]

Prepend x (y ': ys) = (x `AppendSymbol` y) ': Prepend x ys

Combine

union :: IsSet (a :<|> b) => Set a -> Set b -> Set (a :<|> b)

The union of two sets

difference :: IsSet (Difference a b) => Set a -> Set b -> Set (Difference a b)

Difference of two Sets.

type family Difference (a :: [Symbol]) (b :: [Symbol]) :: [Symbol] where ...

Type family encoding of difference

Equations

Difference '[] _ = '[]

Difference xs '[] = xs

Difference xs (y ': ys) = Difference (Delete y xs) ys

intersection :: IsSet (Intersection a b) => Set a -> Set b -> Set (Intersection a b)

The intersection of two Sets.

type family Intersection (a :: [Symbol]) (b :: [Symbol]) :: [Symbol] where ...

Type family encoding of intersection

Equations

Intersection '[] _ = '[]

Intersection (x ': xs) ys = (ElemList x ys :? Intersection xs ys) (x ': Intersection xs ys)

cartesianProduct :: IsSet (CartesianProduct a b) => Set a -> Set b -> Set (CartesianProduct a b)

Calculate the Cartesian product of two Sets. Since there are no tuples, the elements are appended instead of paired in the result.

 cartesianProduct xs ys = fromList $ liftA2 (,) (toList xs) (toList ys)

Example:

 cartesianProduct (fromList [1,2]) (fromList [a,b]) =
   fromList [(1,a), (1,b), (2,a), (2,b)]

type family CartesianProduct (a :: [Symbol]) (b :: [Symbol]) :: [Symbol] where ...

Type family encoding of cartesianProduct

Equations

CartesianProduct '[] _ = '[]

CartesianProduct (x ': xs) ys = Prepend x ys :<|> CartesianProduct xs ys

filterLT :: IsSet (FilterLT a as) => proxy (a :: Symbol) -> Set as -> Set (FilterLT a as)

Only retain elements less than the given one

type family FilterLT (a :: Symbol) (as :: [Symbol]) :: [Symbol] where ...

Type family encoding of filterLT

Equations

FilterLT _ '[] = '[]

FilterLT x (y ': ys) = ((CmpSymbol x y == GT) :? '[]) (y ': FilterLT x ys)

filterGT :: IsSet (FilterGT a as) => proxy (a :: Symbol) -> Set as -> Set (FilterGT a as)

Only retain elements greater than the given one

type family FilterGT (a :: Symbol) (as :: [Symbol]) :: [Symbol] where ...

Type family encoding of filterGT

Equations

FilterGT _ '[] = '[]

FilterGT x (y ': ys) = ((CmpSymbol x y == LT) :? FilterGT x ys) (y ': ys)

split :: (IsSet (FilterLT a as), IsSet (FilterGT a as)) => proxy (a :: Symbol) -> Set as -> (Set (FilterLT a as), Set (FilterGT a as))

The expression (split x set) is a pair (set1,set2) where set1 comprises the elements of set less than x and set2 comprises the elements of set greater than x

splitMember :: (IsSet (FilterLT a as), Reifies (ElemList a as) Bool, IsSet (FilterGT a as)) => proxy (a :: Symbol) -> Set as -> (Set (FilterLT a as), Bool, Set (FilterGT a as))

Performs a split but also returns whether the pivot element was found in the original set.

Indexed

lookupIndex :: Reifies (LookupIndex a as) (Maybe Natural) => proxy (a :: Symbol) -> Set as -> Maybe Natural

Lookup the index of an element, which is its zero-based index in the sorted sequence of elements. The index is a number from 0 up to, but not including, the size of the set.

 isJust   (lookupIndex 2 (fromList [5,3])) == False
 fromJust (lookupIndex 3 (fromList [5,3])) == 0
 fromJust (lookupIndex 5 (fromList [5,3])) == 1
 isJust   (lookupIndex 6 (fromList [5,3])) == False

lookupIndexProxy :: proxy (a :: Symbol) -> Set as -> Proxy (LookupIndex a as :: Maybe Nat)

Type-level lookupIndex

type family LookupIndex (a :: Symbol) (as :: [Symbol]) :: Maybe Nat where ...

Type family encoding of lookupIndex

Equations

LookupIndex _ '[] = Nothing

LookupIndex x (y ': ys) = ((x == y) :? MaybeAddOne (LookupIndex x ys)) (Just 0)

type family MaybeAddOne (a :: Maybe Nat) = (result :: Maybe Nat) where ...

Equivalent to fmap (+ 1)

Equations

MaybeAddOne Nothing = Nothing

MaybeAddOne (Just x) = Just (x + 1)

findIndex :: (KnownNat (FindIndex a as), ElemList a as ~ True) => proxy (a :: Symbol) -> Set as -> Natural

Return the index of an element, which is its zero-based index in the sorted sequence of elements. The index is a number from 0 up to, but not including, the size of the set.

 findIndex 2 (fromList [5,3])    Error: element is not in the set
 findIndex 3 (fromList [5,3]) == 0
 findIndex 5 (fromList [5,3]) == 1
 findIndex 6 (fromList [5,3])    Error: element is not in the set

findIndexProxy :: (KnownNat (FindIndex a as), ElemList a as ~ True) => proxy (a :: Symbol) -> Set as -> Proxy (FindIndex a as)

Type-level findIndex

type family FindIndex (a :: Symbol) (as :: [Symbol]) :: Nat where ...

Type family encoding of findIndex

Equations

FindIndex x (y ': ys) = ((x == y) :? (1 + FindIndex x ys)) 0

FindIndex _ _ = TypeError (Text "Impossible case, did you drop the ElemList constraint?")

elemAt :: ((i < Size a) ~ True, KnownSymbol (ElemAt i a)) => proxy (i :: Nat) -> Set a -> String

Retrieve an element by its index, i.e. by its zero-based index in the sorted sequence of elements. If the index is out of range (less than zero, greater or equal to size of the set), error is called.

 elemAt 0 (fromList [5,3]) == 3
 elemAt 1 (fromList [5,3]) == 5
 elemAt 2 (fromList [5,3])    Error: index out of range

elemAtProxy :: (i < Size a) ~ True => proxy (i :: Nat) -> Set a -> Proxy (ElemAt i a :: Symbol)

Type-level elemAt

type family (a :: Nat) < (b :: Nat) :: Bool where ...

Type family representing less than

Equations

x < y = (x <=? y) && Not (x == y)

type family Not (a :: Bool) = (res :: Bool) | res -> a where ...

Type level not

Equations

Not False = True

Not True = False

type family (a :: k) /= (b :: k) :: Bool where ...

Type level (/=)

Equations

x /= y = Not (x == y)

type family (a :: Bool) && (b :: Bool) where ...

Tyoe level (&&)

Equations

True && True = True

_ && _ = False

type family ElemAt (i :: Nat) (a :: [Symbol]) :: Symbol where ...

Type family encoding of elemAt

Equations

ElemAt 0 (x ': _) = x

ElemAt i (_ ': xs) = ElemAt (i - 1) xs

ElemAt _ _ = TypeError (Text "Impossible case:" :$$: Text "Unless you applied ElemAt without the cooresponding `i < Size a` requirement")

deleteAt :: IsSet (DeleteAt i a) => proxy (i :: Nat) -> Set a -> Set (DeleteAt i a)

Delete the element at index, i.e. by its zero-based index in the sorted sequence of elements. If the index is out of range (less than zero, greater or equal to size of the set), error is called.

 deleteAt 0    (fromList [5,3]) == singleton 5
 deleteAt 1    (fromList [5,3]) == singleton 3
 deleteAt 2    (fromList [5,3])    Error: index out of range
 deleteAt (-1) (fromList [5,3])    Error: index out of range

type family DeleteAt (i :: Nat) (a :: [Symbol]) :: [Symbol] where ...

Type family encoding of deleteAt

Equations

DeleteAt _ '[] = '[]

DeleteAt 0 (_ ': xs) = xs

DeleteAt i (x ': xs) = x ': DeleteAt (i - 1) xs

take :: IsSet (Take i a) => proxy (i :: Nat) -> Set a -> Set (Take i a)

Take a given number of elements in order, beginning with the smallest ones.

type family Take (i :: Nat) (a :: [Symbol]) :: [Symbol] where ...

Type family encoding of take

Equations

Take 0 _ = '[]

Take _ '[] = '[]

Take i (x ': xs) = x ': Take (i - 1) xs

drop :: IsSet (Drop i a) => proxt (i :: Nat) -> Set a -> Set (Drop i a)

Drop a given number of elements in order, beginning with the smallest ones.

type family Drop (i :: Nat) (a :: [Symbol]) :: [Symbol] where ...

Type family encoding of drop

Equations

Drop 0 xs = xs

Drop _ '[] = '[]

Drop i (_ ': xs) = Drop (i - 1) xs

splitAt :: (IsSet (Take i a), IsSet (Drop i a)) => proxy (i :: Nat) -> Set a -> (Set (Take i a), Set (Drop i a))

Split a set at a particular index.

 splitAt n xs = (take n xs, drop n xs)

lookupMin :: ReifySet a => Set a -> Maybe String

The minimal element of a set.

lookupMax :: ReifySet a => Set a -> Maybe String

The maximal element of a set.

findMin :: ReifySet (a ': as) => Set (a ': as) -> String

The minimal element of a set.

findMax :: ReifySet (a ': as) => Set (a ': as) -> String

The maximal element of a set.

deleteMin :: Set (a ': as) -> Set as

Delete the minimal element. Returns an empty set if the set is empty.

deleteMax :: IsSet (DeleteMax a) => Set a -> Set (DeleteMax a)

Delete the maximal element. Returns an empty set if the set is empty.

type family DeleteMax (a :: [Symbol]) :: [Symbol] where ...

Type family encoding of deleteMax

Equations

DeleteMax '[] = '[]

DeleteMax '[_] = '[]

DeleteMax (_ ': (x ': xs)) = DeleteMax (x ': xs)

deleteFindMin :: Set (a ': as) -> (String, Set as)

Delete and find the minimal element. deleteFindMin set = (findMin set, deleteMin set)

deleteFindMax :: (IsSet (DeleteMax (a ': as)), ReifySet (a ': as)) => Set (a ': as) -> (String, Set (DeleteMax (a ': as)))

Delete and find the maximal element. deleteFindMax set = (findMax set, deleteMax set)

maxView :: (IsSet (DeleteMax a), ReifySet a) => Set a -> Maybe (String, Set (DeleteMax a))

Retrieves the maximal key of the set, and the set stripped of that element, or Nothing if passed an empty set.

minView :: ReifySet (a ': as) => Set (a ': as) -> Maybe (String, Set as)

Retrieves the minimal key of the set, and the set stripped of that element, or Nothing if passed an empty set.

Other helpers

liftReflect2 :: Reifies s c => (a -> b -> proxy s) -> a -> b -> c

Lift a function returning a type that we can reflect

Orphan instances

Reifies False Bool

Instance details

Methods

reflect :: proxy False -> Bool

Reifies True Bool

Instance details

Methods

reflect :: proxy True -> Bool

Reifies s a => Reifies ((,) (Proxy :: Proxy a) (Just s) :: (Proxy a, Maybe k)) (Maybe a)

Instance details

Methods

reflect :: proxy (Proxy0, Just s) -> Maybe a

Reifies s a => Reifies ((,) (Proxy :: Proxy a) (Nothing :: Maybe k) :: (Proxy a, Maybe k)) (Maybe a)

Instance details

Methods

reflect :: proxy (Proxy0, Nothing) -> Maybe a