Data.Type.Set

Documentation

type family AllElemsC (c :: k -> Constraint) (as :: [k]) :: Constraint where ...

The constraint that all of the elements of the given list satisfy the given constraint.

Equations

AllElemsC _ '[] = ()

AllElemsC c (x ': xs) = (c x, AllElemsC c xs)

newtype AllElems (c :: k -> Constraint) (as :: [k])

A newtype-wrapped dictionary of the constraint AllElemsC

Constructors

AllElems

Fields

allElemsDict :: Dict (AllElemsC c as)

allElems :: AllElemsC c as => AllElems c as

Construct an instance of AllElems

emptyAllElems :: forall (c :: k -> Constraint). AllElems c '[]

AllElems is vacuously true for any Constraint on the empty list

headAllElems :: AllElems c (a ': as) -> Dict (c a)

If AllElems c (a ': as)` then c a

tailAllElems :: AllElems c (a ': as) -> AllElems c as

If AllElems c (a ': as)` then AllElems c as

unconsAllElems :: AllElems c (a ': as) -> (Dict (c a), AllElems c as)

headAllElems and tailAllElems

consAllElems :: Dict (c a) -> AllElems c as -> AllElems c (a ': as)

Concatenate an instance of a Constraint onto the front of an AllElems

class ReifiesList (as :: [k]) (a :: *) where

Type-class of types that reify to a list of a given type.

If I remember correctly, the instance ReifiesList [] a prevents a functional dependency from as to a.

Minimal complete definition

reflectList

Methods

reflectList :: proxy as -> [a]

class Compare (a :: k)

Type class of types that may be compared with another of the same kind

Associated Types

type Cmp (a :: k) (b :: k) :: Ordering

type family Inc (as :: [k]) :: Constraint where ...

The Constraint that a type-level list is strictly increasing

Equations

Inc '[] = ()

Inc (x ': xs) = Inc1 x xs

type family Inc1 (a :: k) (as :: [k]) = (result :: Constraint) | result -> a as where ...

The tail of Inc

Equations

Inc1 x '[] = x ~ x

Inc1 x (y ': ys) = (x ~ x, Cmp x y ~ LT, Inc1 y ys)

inductMap :: (c '[] :- d '[]) -> (forall (x :: k) (xs :: [k]). c (x ': xs) :- d (x ': xs)) -> c ys :- d ys

undefined

induceMap :: forall (k :: *) (c :: [k] -> Constraint) (ys :: [k]). Dict (c '[]) -> (forall (x :: k) (xs :: [k]). Dict (c (x ': xs))) -> Dict (c (ys :: [k]))

undefined

tailIncEmpty :: forall a. Inc1 a '[] :- Inc '[]

The tail of Inc entails Inc of []

tailIncCons :: forall (a :: k) (b :: k) (bs :: [k]). Inc1 a (b ': bs) :- Inc (b ': bs)

The tail of I

tailInc1 :: Dict (Inc1 a as) -> Dict (Inc as)

In theory, we could induce on tailIncEmpty and tailIncCons...

incToInc1 :: Dict (Inc (a ': as)) -> Dict (Inc1 a as)

Convert Inc of a non-empty type-level list to Inc1

newtype Set (as :: [k])

Singleton Sets, represented on the type-level by type-level strictly increasing lists

Constructors

Set

Fields

setDict :: Dict (Inc as)

set :: Inc as => Set as

If as is strictly increasing, then it represents a valid Set of values

head :: Set (a ': as) -> Proxy a

Get the first value in a non-empty Set

tail :: Set (a ': as) -> Set as

Get the tail of a non-empty Set

type family MinOfC (a :: k) (as :: [k]) :: Constraint where ...

The Constraint that an element of a type-level list is its minimum.

Equations

MinOfC _ '[] = ()

MinOfC x (y ': ys) = (Cmp x y ~ LT, MinOfC x ys)

newtype MinOf (a :: k) (as :: [k])

The singleton representing a proof that an element of a type-level list is its minimum.

Constructors

MinOf

Fields

minOfDict :: Dict (MinOfC a as)

minOf :: MinOfC a as => MinOf a as

MinOf may be instantiated whenever MinOfC holds

type family MaxOfC (a :: k) (as :: [k]) :: Constraint where ...

See MinOfC

Equations

MaxOfC _ '[] = ()

MaxOfC x (y ': ys) = (Cmp x y ~ GT, MaxOfC x ys)

newtype MaxOf (a :: k) (as :: [k])

See MinOf

Constructors

MaxOf

Fields

maxOfDict :: Dict (MaxOfC a as)

maxOf :: MaxOfC a as => MaxOf a as

See minOf

type family Postpend (a :: k) (as :: [k]) :: [k] where ...

Postpend an element to a type-level list

Equations

Postpend x '[] = '[x]

Postpend x (y ': ys) = y ': Postpend x ys

type family Insert (a :: k) (as :: [k]) :: [k] where ...

(Incomplete) Insert an element into a strictly increasing type-level list

Equations

Insert x '[] = '[x]

type family Delete (a :: k) (as :: [k]) :: [k] where ...

(Incomplete) Delete an element from a type-level list

Equations

Delete _ '[] = '[]